refactor: rename mock_grafana → mock_grafana_backend
Made-with: Cursor
