+++ tests/synthetic/mock_grafana_backend/__init__.py
@@ -1,5 +1,5 @@
 """Mock Grafana formatter functions and backends for synthetic testing."""
 
-from tests.synthetic.mock_grafana.backend import FixtureGrafanaBackend, GrafanaBackend
+from tests.synthetic.mock_grafana_backend.backend import FixtureGrafanaBackend, GrafanaBackend
 
 __all__ = ["FixtureGrafanaBackend", "GrafanaBackend"]
